Rename files of classes
