+++ UI/pages/CartPage.py
@@ -1,6 +1,6 @@
 from time import sleep
 
-from .base_page import BasePage
+from .BasePage import BasePage
 from .locators import CartPageLocators
