+++ UI/pages/EditProfilePage.py
@@ -1,4 +1,4 @@
-from .base_page import BasePage
+from .BasePage import BasePage
 from .locators import EditProfilePageLocators
