+++ UI/pages/LoginPage.py
@@ -2,7 +2,7 @@ from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as ec
 from selenium.common.exceptions import TimeoutException
 
-from .base_page import BasePage
+from .BasePage import BasePage
 from .locators import BasePageLocators, LoginPageLocators
