+++ UI/pages/ProductPage.py
@@ -3,7 +3,7 @@ import re
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as ec
 
-from .base_page import BasePage
+from .BasePage import BasePage
 from .locators import ProductPageLocators, HeaderLocators
