+++ UI/pages/ProfilePage.py
@@ -1,6 +1,6 @@
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as ec
-from .base_page import BasePage
+from .BasePage import BasePage
 from .locators import ProfilePageLocators
