+++ UI/pages/locators.py
@@ -57,7 +57,7 @@ class BasePageLocators:
 
 
 class CartPageLocators:
-    AMOUNT = (By.XPATH, '/html/body/main/div/div[1]/div[1]/div[2]/div/div[1]/span')
+    AMOUNT = (By.XPATH, '//button[@id="min-btn"]/following-sibling::span')
     REMOVE_BUTTON = (By.ID, 'remove-all-btn')
     CONTINUE_SHOPPING_BUTTON = (By.ID, 'continue-btn')
     EMPTY_CART_MESSAGE = (By.XPATH, "//*[contains(text(), 'Your cart is empty')]")
@@ -65,10 +65,9 @@ class CartPageLocators:
     MINUS_2_BUTTON = (By.XPATH, '(//button[@id="min-btn"])[2]')
     PLUS_BUTTON = (By.XPATH, '(//button[@id="plus-btn"])[1]')
     PLUS_2_BUTTON = (By.XPATH, '(//button[@id="plus-btn"])[2]')
-    PRODUCT_NAME = (By.XPATH, '//div/div[1]/div/div[2]/p[1]')
-    PRODUCT_COST = (By.XPATH, '//div/div[1]/div[1]/div[2]/p[3]')
-    PRODUCT_WEIGHT = (By.XPATH, "//p[@class='font-medium text-placeholder']")
-    PRODUCT_2_COST = (By.XPATH, '//div/div[1]/div[2]/div[2]/p[3]')
+    PRODUCT_NAME = (By.CSS_SELECTOR, '.text-XL')
+    PRODUCT_COST = (By.XPATH, '(//div/div/div[2]/p)[1]')
+    PRODUCT_2_COST = (By.XPATH, '(//div/div/div[2]/p)[2]')
     SUBTOTAL = (By.XPATH, '/html/body/main/div/div[2]/p[2]')
