+++ UI/test_change_email.py
@@ -1,5 +1,5 @@
-from .pages.profile_page import ProfilePage
-from .pages.edit_profile_page import EditProfilePage
+from .pages.ProfilePage import ProfilePage
+from .pages.EditProfilePage import EditProfilePage
 from .set_of_steps import go_to_edit_profile_page
 from .configs import link, new_email_positive
