+++ UI/test_main_page.py
@@ -2,7 +2,7 @@ from allure import step, title, severity, story, severity_level
 import pytest
 # from time import sleep
 
-from .pages.base_page import BasePage
+from .pages.BasePage import BasePage
 from .configs import link
