+++ UI/test_product_page.py
@@ -1,10 +1,10 @@
 from allure import step, title, severity, story, severity_level
 from time import sleep
 
-from .pages.base_page import BasePage
-from .pages.favorites_page import FavoritesPage
-from .pages.product_page import ProductPage
-from .pages.cart_page import CartPage
+from .pages.BasePage import BasePage
+from .pages.FavoritesPage import FavoritesPage
+from .pages.ProductPage import ProductPage
+from .pages.CartPage import CartPage
 from .set_of_steps import login_user, remove_products_from_cart_and_favorites
 from .configs import link
